A lot more clean up and fixes
