+++ app/src/main/java/be/ugent/zeus/hydra/service/urgent/MusicBinder.java
@@ -7,6 +7,7 @@ import android.os.Binder;
  *
  * @author Niko Strijbol
  */
+@Deprecated
 public class MusicBinder extends Binder {
 
     private final MusicService service;
