+++ app/src/main/java/be/ugent/zeus/hydra/service/urgent/MusicCallback.java
@@ -4,6 +4,7 @@ package be.ugent.zeus.hydra.service.urgent;
  *
  * @author Niko Strijbol
  */
+@Deprecated
 public interface MusicCallback {
     void onPermissionRequired(int requestCode, String permission, String rationale);
     void onLoading();
