--- app/src/main/res/layout/fragment_media_control.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:gravity="center_horizontal">
-
-    <ImageView
-        android:id="@+id/albumImage"
-        android:layout_width="@dimen/media_control_image"
-        android:layout_height="@dimen/media_control_image"
-        android:contentDescription="@string/urgent_album_art"
-        app:srcCompat="@drawable/ic_album" />
-
-    <TextView
-        android:id="@+id/titleText"
-        style="?android:attr/textAppearanceMedium"
-        android:textStyle="bold"
-        android:ellipsize="marquee"
-        tools:text="The Smashing Pumpkins"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center_horizontal" />
-
-    <TextView
-        android:id="@+id/artistText"
-        style="?android:attr/textAppearanceMedium"
-        android:textStyle="normal"
-        android:ellipsize="marquee"
-        tools:text="The End is the Beginning is the End (Soundtrack Version)"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center_horizontal" />
-
-    <ImageButton
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        style="?android:attr/borderlessButtonStyle"
-        android:id="@+id/playPauseButton"
-        app:srcCompat="@drawable/ic_play_arrow_24dp" />
-
-</LinearLayout>
