--- app/src/main/res/layout/fragment_urgent.xml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<android.support.v4.widget.NestedScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:paddingTop="@dimen/activity_horizontal_margin">
-
-        <LinearLayout
-            android:id="@+id/urgent_play_wrapper"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-            <ImageView
-                android:id="@+id/albumImage"
-                android:layout_gravity="center_horizontal"
-                android:layout_width="@dimen/media_control_image"
-                android:layout_height="@dimen/media_control_image"
-                android:contentDescription="@string/urgent_album_art"
-                app:srcCompat="@drawable/logo_urgent" />
-
-            <TextView
-                style="?android:attr/textAppearanceMedium"
-                android:textStyle="bold"
-                android:ellipsize="marquee"
-                android:text="@string/urgent_fm"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:gravity="center_horizontal"
-                android:paddingTop="@dimen/content_spacing"
-                android:paddingBottom="@dimen/content_spacing" />
-
-            <ProgressBar
-                android:layout_gravity="center"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginBottom="@dimen/activity_horizontal_margin" />
-
-        </LinearLayout>
-
-        <FrameLayout
-            android:id="@+id/urgent_fragment_wrapper"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content" />
-
-    </LinearLayout>
-</android.support.v4.widget.NestedScrollView>
